:recycle: Refactor subcommands to not require being a package
